roki: added about command, moved some commands around, fixed minor issues in command data, added extra padding in command command
