+++ Roki.Core/Modules/Utility/InfoCommands.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Discord;
@@ -7,6 +6,7 @@ using Discord.Commands;
 using Discord.WebSocket;
 using Roki.Common.Attributes;
 using Roki.Extensions;
+using Roki.Services;
 
 namespace Roki.Modules.Utility
 {
@@ -16,10 +16,47 @@ namespace Roki.Modules.Utility
         public class InfoCommands : RokiSubmodule
         {
             private readonly DiscordSocketClient _client;
+            private readonly IStatsService _stats;
+            private readonly IRokiConfig _config;
 
-            public InfoCommands(DiscordSocketClient client)
+            public InfoCommands(DiscordSocketClient client, IStatsService stats, IRokiConfig config)
             {
                 _client = client;
+                _stats = stats;
+                _config = config;
+            }
+
+            [RokiCommand, Description, Usage, Aliases]
+            public async Task About()
+            {
+                await Context.Channel.EmbedAsync(new EmbedBuilder().WithOkColor()
+                        .WithAuthor(_stats.Author)
+                        .WithTitle($"Roki v{StatsService.BotVersion}")
+                        .WithThumbnailUrl(Context.Client.CurrentUser.GetAvatarUrl())
+                        .WithDescription($"Roki is an general purpose discord bot, made by {_stats.Author}. Project is open source on [GitHub](https://github.com/snoww/roki)\n" +
+                                         $"Written in C# using Discord.Net v{DiscordConfig.Version}. Inspired by NadekoBot and yagpdb." +
+                                         "Roki has many features, including but not limited to: XP system, Currency system, Event system, Jeopardy!, and many more." +
+                                         "If you have any questions/issues feel free to DM Snow#7777"))
+                    .ConfigureAwait(false);
+
+            }
+            
+            [RokiCommand, Description, Usage, Aliases]
+            public async Task Stats()
+            {
+                var ownerId = string.Join("\n", _config.OwnerIds);
+                if (string.IsNullOrWhiteSpace(ownerId)) ownerId = "-";
+
+                await Context.Channel.EmbedAsync(
+                    new EmbedBuilder().WithOkColor()
+                        .WithAuthor($"Roki v{StatsService.BotVersion}", Context.Client.CurrentUser.GetAvatarUrl())
+                        .AddField("Bot ID", _client.CurrentUser.Id, true)
+                        .AddField("Owner ID", ownerId, true)
+                        .AddField("Commands ran", _stats.CommandsRan, true)
+                        .AddField("Messages", _stats.MessageCounter, true)
+                        .AddField("Memory", $"{_stats.Heap} MB", true)
+                        .AddField("Uptime", _stats.GetUptimeString("\n"), true)
+                        .AddField("Presence", $"{_stats.TextChannels} Text Channels\n{_stats.VoiceChannels} Voice Channels", true));
             }
 
             [RokiCommand, Usage, Description, Aliases, RequireContext(ContextType.Guild)]
@@ -150,22 +187,6 @@ namespace Roki.Modules.Utility
                     .AddField("Avatar Url", avatarUrl, true);
 
                 await Context.Channel.EmbedAsync(embed, Context.User.Mention).ConfigureAwait(false);
-            }
-
-            [RokiCommand, Description, Usage, Aliases]
-            public async Task Ping()
-            {
-                var sw = Stopwatch.StartNew();
-                var msg = await Context.Channel.SendMessageAsync("🏓").ConfigureAwait(false);
-                sw.Stop();
-                await msg.DeleteAsync().ConfigureAwait(false);
-
-                var embed = new EmbedBuilder();
-                embed.WithOkColor()
-                    .WithAuthor("Pong! 🏓")
-                    .WithDescription($"Currently {(int) sw.Elapsed.TotalMilliseconds}ms");
-
-                await Context.Channel.EmbedAsync(embed).ConfigureAwait(false);
             }
         }
     }
