+++ Roki.Core/Modules/Utility/Utility.cs
@@ -1,47 +1,16 @@
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
-using Discord.WebSocket;
 using Roki.Common.Attributes;
 using Roki.Extensions;
 using Roki.Modules.Utility.Services;
-using Roki.Services;
 
 namespace Roki.Modules.Utility
 {
     public partial class Utility : RokiTopLevelModule<UtilityService>
     {
-        private readonly DiscordSocketClient _client;
-        private readonly IRokiConfig _config;
-        private readonly IStatsService _stats;
-
-        public Utility(DiscordSocketClient client, IStatsService stats, IRokiConfig config)
-        {
-            _client = client;
-            _stats = stats;
-            _config = config;
-        }
-
-        [RokiCommand, Description, Usage, Aliases]
-        public async Task Stats()
-        {
-            var ownerId = string.Join("\n", _config.OwnerIds);
-            if (string.IsNullOrWhiteSpace(ownerId)) ownerId = "-";
-
-            await Context.Channel.EmbedAsync(
-                new EmbedBuilder().WithOkColor()
-                    .WithAuthor(eab => eab.WithName($"Roki v{StatsService.BotVersion}").WithIconUrl(Context.Client.CurrentUser.GetAvatarUrl()))
-                    .AddField("Author", _stats.Author, true)
-                    .AddField("Bot ID", _client.CurrentUser.Id.ToString(), true)
-                    .AddField("Owner ID", ownerId, true)
-                    .AddField("Commands ran", _stats.CommandsRan.ToString(), true)
-                    .AddField("Messages", _stats.MessageCounter, true)
-                    .AddField("Memory", $"{_stats.Heap} MB", true)
-                    .AddField("Uptime", _stats.GetUptimeString("\n"), true)
-                    .AddField("Presence", $"{_stats.TextChannels} Text Channels\n{_stats.VoiceChannels} Voice Channels", true));
-        }
-
         [RokiCommand, Description, Usage, Aliases, RequireContext(ContextType.Guild)]
         public async Task Pins()
         {
@@ -80,6 +49,22 @@ namespace Roki.Modules.Utility
             var uwuize = Service.Uwulate(message);
             await Context.Channel.SendMessageAsync(uwuize).ConfigureAwait(false);
         }
+        
+        [RokiCommand, Description, Usage, Aliases]
+        public async Task Ping()
+        {
+            var sw = Stopwatch.StartNew();
+            var msg = await Context.Channel.SendMessageAsync("🏓").ConfigureAwait(false);
+            sw.Stop();
+            await msg.DeleteAsync().ConfigureAwait(false);
+
+            var embed = new EmbedBuilder();
+            embed.WithOkColor()
+                .WithAuthor("Pong! 🏓")
+                .WithDescription($"Currently {(int) sw.Elapsed.TotalMilliseconds} ms");
+
+            await Context.Channel.EmbedAsync(embed).ConfigureAwait(false);
+        }
 
         [RokiCommand, Description, Usage, Aliases]
         [RequireContext(ContextType.Guild)]
@@ -100,7 +85,7 @@ namespace Roki.Modules.Utility
             if (string.IsNullOrWhiteSpace(message))
                 return;
             
-            await Context.Channel.SendMessageAsync(Context.Message.Content).ConfigureAwait(false);
+            await Context.Channel.SendMessageAsync(Format.Code(message)).ConfigureAwait(false);
         }
     }
 }
